Backend/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astroid==1.6.0
Brlapi==0.6.6
certifi==2020.4.5.1
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.6
django-crispy-forms==1.9.1
django-environ==0.4.5
django-filter==2.2.0
django-nested-admin==3.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
isort==4.3.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
logilab-common==1.4.1
louis==3.5.0
macaroonbakery==1.1.3
Markdown==3.2.1
mccabe==0.6.1
mysqlclient==1.4.6
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==7.1.2
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pylint==1.8.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-monkey-business==1.0.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
virtualenv==20.0.20
wadllib==1.3.2
wrapt==1.9.0
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
